apps/api/pyproject.toml
[project]
name = "ghitdesk-api"
version = "0.1.0"
description = "Backend multi-tenant do GhitDesk (Django + DRF + Channels + Celery)."
requires-python = ">=3.12"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15",
  "django-tenants>=3.6",
  "channels[daphne]>=4.1",
  "channels-redis>=4.2",
  "psycopg[binary]>=3.2",
  "dj-database-url>=2.2",
  "django-environ>=0.11",
  "django-redis>=5.4",
  "celery[rabbitmq,redis]>=5.4",
  "drf-spectacular>=0.27",
  "structlog>=24.1",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-instrumentation-django>=0.46",
  "opentelemetry-instrumentation-celery>=0.46",
  "python-json-logger>=2.0",
  "pydantic>=2.8",
]

[tool.uv]
index-url = "https://pypi.org/simple"
